Add code to not cleanup the server if there are other connections to it.

+++ redislite/client.py
@@ -67,35 +67,63 @@ class RedisMixin(object):
         :return:
         """
 
-        if not self.redis_dir:  # pragma: no cover
-            return
+        #if not self.redis_dir:  # pragma: no cover
+        #    return
 
-        if self.running:
+        if self.pid:
+            logger.debug('Connection count: %s', self._connection_count())
+            if self._connection_count() <= 1:
+                logger.debug(
+                    'Last client using the connection, shutting down the redis '
+                    'connection on socket: %s',
+                    self.socket_file
+                )
+                # noinspection PyUnresolvedReferences
+                logger.debug(
+                    'Shutting down redis server with pid of %d' % self.pid)
+                self.shutdown()
+                self.socket_file = None
+
+                if self.pidfile and os.path.exists(self.pidfile):   # pragma: no cover
+                    # noinspection PyTypeChecker
+                    pid = int(open(self.pidfile).read())
+                    os.kill(pid, signal.SIGKILL)
+
+                if self.redis_dir and os.path.isdir(self.redis_dir):  # pragma: no cover
+                    shutil.rmtree(self.redis_dir)
+
+                if self.cleanupregistry and os.path.exists(
+                        self.settingregistryfile
+                ):
+                    os.remove(self.settingregistryfile)
+                    self.settingregistryfile = None
+            else:
+                logger.debug(
+                    'Other clients are still connected to the redis server, '
+                    'not shutting down the connection on socket: %s',
+                    self.socket_file
+                )
+                self.connection_pool.disconnect()
+        else:
             logger.debug(
-                'Shutting down the redis connection on socket: %s',
+                'Redis is not running on socket %s, skipping cleanup',
                 self.socket_file
-            )
-            # noinspection PyUnresolvedReferences
-            logger.debug(
-                'Shutting down redis server with pid of %d' % self.pid)
-            self.shutdown()
-            self.socket_file = None
-
-        if self.pidfile and os.path.exists(self.pidfile):   # pragma: no cover
-            # noinspection PyTypeChecker
-            pid = int(open(self.pidfile).read())
-            os.kill(pid, signal.SIGKILL)
-
-        if os.path.isdir(self.redis_dir):  # pragma: no cover
-            shutil.rmtree(self.redis_dir)
-
-        if self.cleanupregistry and os.path.exists(self.settingregistryfile):
-            os.remove(self.settingregistryfile)
-            self.settingregistryfile = None
+            )  # pragma: no cover
 
         self.running = False
         self.redis_dir = None
         self.pidfile = None
+
+    def _connection_count(self):
+        if not self._is_redis_running():  # pragma: no cover
+            return 0
+        active_connections = 0
+        for client in self.client_list():
+            flags = client.get('flags', '')
+            flags = flags.upper()
+            if 'U' in flags or 'N' in flags:
+                active_connections += 1
+        return active_connections
 
     def _create_redis_directory_tree(self):
         """
@@ -182,7 +210,10 @@ class RedisMixin(object):
                 pid = fh.read().strip()   # NOQA
                 pid = int(pid)
                 if pid:  # pragma: no cover
-                    p = psutil.Process(pid)
+                    try:
+                        p = psutil.Process(pid)
+                    except psutil.NoSuchProcess:
+                        return False
                     if not p.is_running():
                         return False
                 else:  # pragma: no cover
@@ -301,6 +332,9 @@ class RedisMixin(object):
         logger.debug('Calling binding with %s, %s', args, kwargs)
         super(RedisMixin, self).__init__(*args, **kwargs)  # pragma: no cover
 
+        logger.debug("Pinging the server to ensure we're connected")
+        self.ping()
+
     def __del__(self):
         self._cleanup()  # pragma: no cover
 
@@ -326,7 +360,16 @@ class RedisMixin(object):
         """
         if self.pidfile and os.path.exists(self.pidfile):
             with open(self.pidfile) as fh:
-                pid = fh.read().strip()
+                pid = int(fh.read().strip())
+                if pid:  # pragma: no cover
+                    try:
+                        p = psutil.Process(pid)
+                    except psutil.NoSuchProcess:
+                        return 0
+                    if not p.is_running():
+                        return pid
+                else:  # pragma: no cover
+                    return 0
             return int(pid)
         return 0  # pragma: no cover
